+++ crates/cli/src/commands/outdated.rs
@@ -31,9 +31,6 @@ pub struct OutdatedArgs {
         help = "Update and write the versions to their respective configuration"
     )]
     update: bool,
-
-    #[arg(long, help = "Avoid and force confirm prompts", env = "PROTO_YES")]
-    yes: bool,
 }
 
 #[derive(Serialize)]
@@ -248,7 +245,7 @@ pub async fn outdated(session: ProtoSession, args: OutdatedArgs) -> AppResult {
         return Ok(None);
     }
 
-    let skip_prompts = session.skip_prompts(args.yes);
+    let skip_prompts = session.should_skip_prompts();
     let mut confirmed = false;
 
     if !skip_prompts {