+++ crates/cli/src/main.rs
@@ -53,7 +53,7 @@ async fn main() -> MainResult {
         dump_trace: cli.dump && !matches!(cli.command, Commands::Run { .. }),
         filter_modules: get_tracing_modules(),
         log_env: "PROTO_APP_LOG".into(),
-        show_spans: cli.log.as_ref().is_some_and(|level| level.is_verbose()),
+        show_spans: cli.log.is_verbose(),
         // test_env: "PROTO_TEST".into(),
         ..TracingOptions::default()
     });